% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ckmeans.1d.dp.mod.R
\name{Ckmeans.1d.dp.mod}
\alias{Ckmeans.1d.dp.mod}
\title{Ckmeans.1d.dp.mod}
\usage{
Ckmeans.1d.dp.mod(x, z, w, k = 2)
}
\arguments{
\item{x}{vector of one-dimensional values to cluster}

\item{z}{list of vectors used to calculate variance objective (one vector per x value)}

\item{w}{list of vectors representing cell weights (one vector per x value)}

\item{k}{number of clusters}
}
\value{
A list of results containing:
\itemize{
  \item \code{cluster} optimised cluster indices at varying number of clusters
  \item \code{centers} cluster centers at varying number of clusters
  \item \code{size} the size of each cluster for varying number of clusters change
}
}
\description{
Modified one-dimensional K-means function.
}
\details{
Modified function from the Ckmeans.1d.dp R package (attributed to Haizhou Wang and Joe Song).

Segments the one-dimensional vector x into clusters without breaking the continuity
of x values such that only neighbouring values can be grouped together.

Each x value has an associate vector of weights and z values which are used in the
calculation of the weighted sum of squares objective.  The clusters are then optimised
by minimising the weighted sum of squares (according to the vector of weights and z)
such that the continuity of x values is preserved.
}
